Merge old and new codebase
--

The repo will no longer compile properly in it's current state. The plan is to get the code to at least compile, and then to refine it by properly adding Docker support, and using an environment variable-based configuration system instead of storing secrets in XML data.

Another goal is to re-write the database code. I don't like the implementation that I wrote previously, and it's only here to serve as a reference to how I want to structure the data.